rename app, create new app file
